added little play functionality, change password, routes protected

+++ controllers/videoController.js
@@ -35,9 +35,12 @@ export const postUpload = async(req, res) => {
             {
                 fileUrl:path,
                 title,
-                description
+                description,
+                creator: req.user.id
             }
         );
+        req.user.videos.push(newVideo.id);
+        req.user.save();
         console.log(newVideo);
         //res.render("upload", { pageTitle : "Upload"});
         res.redirect(routes.videoDetail(newVideo.id));
@@ -47,7 +50,7 @@ export const postUpload = async(req, res) => {
 export const videoDetail = async(req, res) => {
     const { params: {id}} = req;
     try{
-        const video = await Video.findById(id)
+        const video = await Video.findById(id).populate('creator');
         res.render("videoDetail", { pageTitle : video.title, video});
     }catch(error){
 
@@ -56,16 +59,20 @@ export const videoDetail = async(req, res) => {
 }
 
 export const getEditVideo = async(req, res) => {
-    const { params: {id}} = req;
-    try{
+    const {
+        params: { id }
+      } = req;
+      try {
         const video = await Video.findById(id);
-        res.render("editVideo", { pageTitle : `Edit ${video.title}`, video});
-    }catch(error){
-        res.redirect(routes.home); 
-    }
-
-    res.render("editVideo", { pageTitle : "Edit Video"});
-}
+        if (String(video.creator) !== req.user.id) {
+          throw Error();
+        } else {
+          res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        }
+      } catch (error) {
+        res.redirect(routes.home);
+      }
+    };
 
 export const postEditVideo = async(req, res) => {
     const {params:{id}, body : {title, description}} = req;
@@ -80,12 +87,19 @@ export const postEditVideo = async(req, res) => {
     res.render("editVideo", { pageTitle : "Edit Video"});
 }
 
-export const deleteVideo = async(req, res) => {
-    const {params:{id}} = req;
-    try{
-        await Video.findOneAndRemove({_id: id})
-    }catch(error){
-        console.log(error)
+export const deleteVideo = async (req, res) => {
+    const {
+      params: { id }
+    } = req;
+    try {
+      const video = await Video.findById(id);
+      if (String(video.creator) !== req.user.id) {
+        throw Error();
+      } else {
+        await Video.findOneAndRemove({ _id: id });
+      }
+    } catch (error) {
+      console.log(error);
     }
-    res.redirect(routes.home)
-}
+    res.redirect(routes.home);
+  };
